# api/openapi.yaml
openapi: 3.0.3
info:
  title: Stampede API
  version: 1.0.0
  description: |
    The public REST API of a Stampede server. The web UI, the CLI and CI
    integrations all use this API.

    Authenticate with a session cookie (from `POST /auth/login`) or an API
    token sent as `Authorization: Bearer stp_...`. Requests that use the
    session cookie and change state must also send `X-Stampede-CSRF: 1`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
security:
  - bearer: []
  - session: []
...
paths:
...
  /setup:
    post:
      tags: [system]
      operationId: setup
      summary: Create the first organisation and owner account
      description: Only allowed while the server has no users.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "201":
          description: Owner created and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Signed in; sets the session cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooMany" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: API token, `stp_...`
    session:
      type: apiKey
      in: cookie
      name: stampede_session
...
  responses:
    Unauthorized:
      description: Not signed in or token invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: Conflicts with the current state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: The request is well-formed but invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooMany:
      description: Too many attempts
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: not_found }
            message: { type: string }
            details:
              type: array
              items: { type: string }
...
    Role:
      type: string
      enum: [owner, admin, editor, runner, viewer]
    SetupRequest:
      type: object
      required: [organisation, name, email, password]
      properties:
        organisation: { type: string, minLength: 1, maxLength: 100 }
        name: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 10 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    Session:
      type: object
      required: [user, expiresAt]
      properties:
        user: { $ref: "#/components/schemas/Me" }
        expiresAt: { type: string, format: date-time }
    Me:
      type: object
      required: [id, email, name, role, orgId, orgName]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        orgId: { type: string, format: uuid }
        orgName: { type: string }
...