# internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package db

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type AiJob struct {
	ID                 uuid.UUID
	OrgID              uuid.UUID
	ProjectID          uuid.UUID
	ProviderID         *uuid.UUID
	ProviderKind       string
	Model              string
	Status             string
	Stage              string
	Round              int32
	TargetID           *uuid.UUID
	ScenarioID         *uuid.UUID
	DryRun             bool
	Inputs             json.RawMessage
	Yaml               string
	Diff               string
	Result             json.RawMessage
	Error              string
	InputTokens        int64
	OutputTokens       int64
	CreatedBy          *uuid.UUID
	CreatedAt          time.Time
	StartedAt          *time.Time
	FinishedAt         *time.Time
	ApprovedAt         *time.Time
	ApprovedScenarioID *uuid.UUID
	ApprovedVersion    *int32
	OwnerReplica       *uuid.UUID
}

type AiProvider struct {
	ID              uuid.UUID
	OrgID           uuid.UUID
	Name            string
	Kind            string
	Model           string
	BaseUrl         string
	Ciphertext      []byte
	WrappedKey      []byte
	KeyID           *string
	MonthlyTokenCap int64
	CreatedBy       *uuid.UUID
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type AiUsage struct {
	ID           uuid.UUID
	OrgID        uuid.UUID
	ProviderID   *uuid.UUID
	Purpose      string
	RunID        *uuid.UUID
	InputTokens  int64
	OutputTokens int64
	CreatedBy    *uuid.UUID
	CreatedAt    time.Time
}

type ApiToken struct {
	ID         uuid.UUID
	OrgID      uuid.UUID
	UserID     uuid.UUID
	Name       string
	Prefix     string
	TokenHash  []byte
	Role       string
	CreatedAt  time.Time
	LastUsedAt *time.Time
	ExpiresAt  *time.Time
}

type AuditLog struct {
	ID      int64
	OrgID   uuid.UUID
	At      time.Time
	Actor   string
	ActorID *uuid.UUID
	Action  string
	Subject string
	Details json.RawMessage
	Ip      string
}

type DriftResult struct {
	ID              uuid.UUID
	ProjectID       uuid.UUID
	ScheduleID      *uuid.UUID
	ScenarioID      uuid.UUID
	ScenarioVersion int32
	TargetID        uuid.UUID
	Status          string
	Error           string
	Broken          []string
	Result          json.RawMessage
	Spec            json.RawMessage
	RepairJobID     *uuid.UUID
	CreatedBy       *uuid.UUID
	CreatedAt       time.Time
}

type Integration struct {
	ID         uuid.UUID
	OrgID      uuid.UUID
	Name       string
	Kind       string
	Url        string
	Ciphertext []byte
	WrappedKey []byte
	KeyID      *string
	CreatedBy  *uuid.UUID
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type Membership struct {
	OrgID     uuid.UUID
	UserID    uuid.UUID
	Role      string
	CreatedAt time.Time
}

type NotificationChannel struct {
	ID               uuid.UUID
	OrgID            uuid.UUID
	Name             string
	Kind             string
	Events           []string
	AllowPrivate     bool
	UrlHint          string
	UrlCiphertext    []byte
	UrlWrappedKey    []byte
	SecretCiphertext []byte
	SecretWrappedKey []byte
	KeyID            string
	CreatedBy        *uuid.UUID
	CreatedAt        time.Time
}

type NotificationDelivery struct {
	ID         int64
	ChannelID  uuid.UUID
	DeliveryID uuid.UUID
	Event      string
	RunID      *uuid.UUID
	Attempt    int32
	Ok         bool
	StatusCode int32
	Error      string
	DurationMs int32
	At         time.Time
}

type Org struct {
	ID        uuid.UUID
	Name      string
	CreatedAt time.Time
}

type OrgCap struct {
	OrgID        uuid.UUID
	MaxRate      *float64
	MaxVus       *int32
	MaxDurationS *int32
	UpdatedBy    *uuid.UUID
	UpdatedAt    time.Time
}

type Project struct {
	ID          uuid.UUID
	OrgID       uuid.UUID
	Name        string
	Slug        string
	Description string
	CreatedAt   time.Time
}

type ProjectRole struct {
	ProjectID uuid.UUID
	UserID    uuid.UUID
	Role      string
	CreatedBy *uuid.UUID
	CreatedAt time.Time
}

type ProjectSetting struct {
	ProjectID     uuid.UUID
	MaxRate       *float64
	MaxVus        *int32
	MaxDurationS  *int32
	RequireDryRun bool
	UpdatedBy     *uuid.UUID
	UpdatedAt     time.Time
}

type Replica struct {
	ID        uuid.UUID
	Addr      string
	Version   string
	StartedAt time.Time
	LastSeen  time.Time
}

type Report struct {
	RunID     uuid.UUID
	Report    json.RawMessage
	CreatedAt time.Time
}

type Run struct {
	ID              uuid.UUID
	ProjectID       uuid.UUID
	ScenarioID      uuid.UUID
	ScenarioVersion int32
	TargetID        uuid.UUID
	Status          string
	Verdict         *string
	StopReason      *string
	Error           *string
	Overrides       json.RawMessage
	Plan            json.RawMessage
	Env             json.RawMessage
	Workers         int32
	Note            string
	Summary         json.RawMessage
	CreatedBy       *uuid.UUID
	CreatedAt       time.Time
	StartedAt       *time.Time
	EndedAt         *time.Time
	OwnerReplica    *uuid.UUID
}

type RunEvent struct {
	ID      int64
	RunID   uuid.UUID
	At      time.Time
	Type    string
	Message string
	Worker  string
	Details json.RawMessage
}

type RunMetric struct {
	RunID      uuid.UUID
	Ts         time.Time
	Interval   int32
	Requests   int64
	Failed     int64
	Rps        float64
	ErrorRate  float64
	P50        float64
	P95        float64
	P99        float64
	Vus        int32
	Planned    float64
	Dropped    int64
	Iterations int64
	SchedLag   float64
	Snapshot   []byte
}

type Scenario struct {
	ID          uuid.UUID
	ProjectID   uuid.UUID
	Name        string
	Description string
	Tags        []string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type ScenarioVersion struct {
	ScenarioID uuid.UUID
	Version    int32
	Yaml       string
	Message    string
	Plan       json.RawMessage
	CreatedBy  *uuid.UUID
	CreatedAt  time.Time
}

type Schedule struct {
	ID             uuid.UUID
	ProjectID      uuid.UUID
	Name           string
	ScenarioID     uuid.UUID
	TargetID       uuid.UUID
	Cron           string
	Timezone       string
	Overrides      json.RawMessage
	Env            json.RawMessage
	Workers        int32
	Enabled        bool
	Note           string
	OwnerID        *uuid.UUID
	CreatedAt      time.Time
	UpdatedAt      time.Time
	NextRunAt      *time.Time
	LastFiredAt    *time.Time
	LastRunID      *uuid.UUID
	LastSkipReason string
	Kind           string
	SpecUrl        string
	LastDriftID    *uuid.UUID
}

type Secret struct {
	ProjectID  uuid.UUID
	Name       string
	Ciphertext []byte
	WrappedKey []byte
	KeyID      string
	UpdatedAt  time.Time
}

type Session struct {
	TokenHash []byte
	UserID    uuid.UUID
	OrgID     uuid.UUID
	CreatedAt time.Time
	ExpiresAt time.Time
	Ip        string
	UserAgent string
}

type Target struct {
	ID                 uuid.UUID
	ProjectID          uuid.UUID
	Name               string
	BaseUrl            string
	Host               string
	Private            bool
	VerificationToken  string
	VerifiedAt         *time.Time
	VerificationMethod *string
	AllowHosts         []string
	MaxRate            *float64
	MaxVus             *int32
	MaxDurationS       *int32
	CreatedAt          time.Time
}

type User struct {
	ID           uuid.UUID
	Email        string
	Name         string
	PasswordHash string
	CreatedAt    time.Time
	LastLoginAt  *time.Time
}